Address review comments from copilot-pull-request-reviewer: enable Kafka receive spans in preview
