+++ instrumentation/kafka/kafka-clients/kafka-clients-0.11/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/kafkaclients/v0_11/KafkaClientDefaultTest.java
@@ -10,8 +10,11 @@ import static io.opentelemetry.instrumentation.testing.util.TelemetryDataUtil.or
 import static io.opentelemetry.sdk.testing.assertj.OpenTelemetryAssertions.assertThat;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.concurrent.TimeUnit.SECONDS;
+import static org.junit.jupiter.api.Assumptions.assumeTrue;
 
 import io.opentelemetry.api.trace.SpanKind;
+import io.opentelemetry.context.Context;
+import io.opentelemetry.context.Scope;
 import io.opentelemetry.instrumentation.kafkaclients.common.v0_11.internal.KafkaClientBaseTest;
 import io.opentelemetry.instrumentation.kafkaclients.common.v0_11.internal.KafkaClientPropagationBaseTest;
 import io.opentelemetry.instrumentation.testing.junit.AgentInstrumentationExtension;
@@ -19,6 +22,7 @@ import io.opentelemetry.instrumentation.testing.junit.InstrumentationExtension;
 import io.opentelemetry.sdk.trace.data.LinkData;
 import io.opentelemetry.sdk.trace.data.SpanData;
 import java.time.Duration;
+import java.util.Iterator;
 import java.util.List;
 import java.util.ListIterator;
 import java.util.concurrent.atomic.AtomicReference;
@@ -187,6 +191,43 @@ class KafkaClientDefaultTest extends KafkaClientPropagationBaseTest {
             trace.hasSpansSatisfyingExactly(
                 span ->
                     span.hasName("poll " + SHARED_TOPIC).hasKind(SpanKind.CLIENT).hasNoParent()));
+  }
+
+  @Test
+  void testAbandonedIteratorDoesNotParentNextProcessSpan() throws Exception {
+    assumeTrue(emitStableMessagingSemconv());
+    producer.send(new ProducerRecord<>(SHARED_TOPIC, "first")).get(5, SECONDS);
+    awaitUntilConsumerIsReady();
+    Iterator<? extends ConsumerRecord<?, ?>> firstIterator = poll(Duration.ofSeconds(5)).iterator();
+    assertThat(firstIterator.hasNext()).isTrue();
+    firstIterator.next();
+
+    try (Scope ignored = Context.root().makeCurrent()) {
+      producer.send(new ProducerRecord<>(SHARED_TOPIC, "second")).get(5, SECONDS);
+    }
+    Iterator<? extends ConsumerRecord<?, ?>> secondIterator =
+        poll(Duration.ofSeconds(5)).iterator();
+    assertThat(secondIterator.hasNext()).isTrue();
+    secondIterator.next();
+    assertThat(secondIterator.hasNext()).isFalse();
+    assertThat(firstIterator.hasNext()).isFalse();
+
+    testing.waitAndAssertSortedTraces(
+        orderByRootSpanKind(SpanKind.PRODUCER, SpanKind.CLIENT),
+        trace ->
+            trace.hasSpansSatisfyingExactly(
+                span -> span.hasName("send " + SHARED_TOPIC).hasNoParent(),
+                span -> span.hasName("process " + SHARED_TOPIC).hasParent(trace.getSpan(0))),
+        trace ->
+            trace.hasSpansSatisfyingExactly(
+                span -> span.hasName("send " + SHARED_TOPIC).hasNoParent(),
+                span -> span.hasName("process " + SHARED_TOPIC).hasParent(trace.getSpan(0))),
+        trace ->
+            trace.hasSpansSatisfyingExactly(
+                span -> span.hasName("poll " + SHARED_TOPIC).hasNoParent()),
+        trace ->
+            trace.hasSpansSatisfyingExactly(
+                span -> span.hasName("poll " + SHARED_TOPIC).hasNoParent()));
   }
 
   @DisplayName("test pass through tombstone")
